+++ src/Workspace.cpp
@@ -820,6 +820,43 @@ std::optional<lsp::Location> WorkspaceFolder::gotoTypeDefinition(const lsp::Type
     return std::nullopt;
 }
 
+lsp::ReferenceResult WorkspaceFolder::references(const lsp::ReferenceParams& params)
+{
+    // TODO: currently we only support searching for a binding at a current position
+    auto moduleName = getModuleName(params.textDocument.uri);
+    auto position = convertPosition(params.position);
+
+    // Run the type checker to ensure we are up to date
+    // TODO: we only need the parse result here - can typechecking be skipped?
+    if (frontend.isDirty(moduleName))
+        frontend.check(moduleName);
+
+    auto sourceModule = frontend.getSourceModule(moduleName);
+    auto module = frontend.moduleResolver.getModule(moduleName);
+    if (!sourceModule || !module)
+        return std::nullopt;
+
+    auto exprOrLocal = Luau::findExprOrLocalAtPosition(*sourceModule, position);
+    Luau::Symbol symbol;
+
+    if (exprOrLocal.getLocal())
+        symbol = exprOrLocal.getLocal();
+    else if (auto exprLocal = exprOrLocal.getExpr()->as<Luau::AstExprLocal>())
+        symbol = exprLocal->local;
+    else
+        return std::nullopt;
+
+    auto references = findSymbolReferences(*sourceModule, symbol);
+    std::vector<lsp::Location> result;
+
+    for (auto& location : references)
+    {
+        result.emplace_back(lsp::Location{params.textDocument.uri, {convertPosition(location.begin), convertPosition(location.end)}});
+    }
+
+    return result;
+}
+
 // std::optional<std::vector<lsp::DocumentSymbol>> WorkspaceFolder::documentSymbol(const lsp::DocumentSymbolParams& params)
 // {
 //     auto moduleName = getModuleName(params.textDocument.uri);